+++ src/spicelib/devices/numd2/nud2load.c
@@ -328,7 +328,7 @@ NUMD2load(GENmodel *inModel, CKTcircuit *ckt)
 	    ckt->CKTtroubleElt = (GENinstance *) inst;
 	    return (E_BADMATRIX);
 	  }
-	  devConverged = TWOdeviceConverged(pDevice);
+	  devConverged = (int)TWOdeviceConverged(pDevice);
 	  pDevice->converged = devConverged;
 
 	  /* extract the current and conductance information */
